requirements.txt
opencv-python==4.10.0.82
psutil==5.9.8
PyAutoGUI==0.9.54
pytesseract==0.3.10
pillow==10.3.0